requirements.txt
python_telegram_bot==12.7
requests==2.23.0
cognitive_face==1.5.0
Pillow==7.1.2
telegram==0.0.1
